[#78]Add Physics Proxy
Adds a physics proxy to the active object.  Can only operate on one
object at a time.  Automatically adds the proxy to the same group, gives
it a proxy material, and checks if proxy already exists for object.
Allows you to select type, which doesn't effect the display but flags it
so that CryEngine will convert it to this shape using UDP.