Allow materials from another nodes
If the material assigned to a node is referencing another node (i.e. generic materials), the material will be assigned as it is already and not added to the .mtl file. If the material is referencing the current node or it's using the new (not assigned to node) material notation, will be added to the .mtl file.