New "Add Material" menu feature
New material can be created for all the selected objects specifing the
material name and the physics to use. The function uses the current
cryexportnode name and take care about the needed index for the new
material. The material is created for the group so selecting meshes in
different groups will create different materials, one per group. The
option is available in the main menu and in the left panel.